CADC-9366 moved auth check to before Job creation, other rework
