+++ luskan/src/main/java/org/opencadc/luskan/LuskanConfig.java
@@ -175,6 +175,7 @@ public class LuskanConfig {
         List<String> allowedGroups = props.getProperty(ALLOWED_GROUP);
         if (allowedGroups.isEmpty()) {
             sb.append("MISSING");
+            ok = false;
         } else {
             for (String allowedGroup : allowedGroups) {
                 sb.append("\n\t").append(ALLOWED_GROUP).append(" - ").append(allowedGroup);
@@ -183,6 +184,7 @@ public class LuskanConfig {
                     sb.append(" OK");
                 } catch (IllegalArgumentException e) {
                     sb.append(" INVALID GroupURI");
+                    ok = false;
                 }
             }
         }
